backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
httpx>=0.28.1,<1.0.0
aiohttp==3.9.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pymongo==4.8.0
motor==3.5.1
pydantic==2.8.2
redis>=4.5.0
aioredis==2.0.1
celery>=5.3.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0
bleach>=6.1.0
aiofiles>=23.2.1
psutil>=5.9.6
openai>=1.3.8
google-genai>=1.0.0
tiktoken>=0.5.2
cryptography>=41.0.8
certifi>=2023.11.17
structlog>=23.2.0
prometheus-client>=0.19.0
aiosmtplib>=3.0.1
Pillow>=10.1.0
python-magic>=0.4.27
pandas>=2.1.4
numpy>=1.25.2
slowapi>=0.1.9
websockets>=12.0
lz4>=4.3.2
pytz>=2023.3
shortuuid>=1.0.1
environs>=10.0.0
email-validator>=2.1.0
phonenumbers>=8.13.25
aiocache>=0.12.2
dramatiq>=1.14.2
pyinstrument>=0.4.0
fastapi-health>=0.4.0
fastapi-limiter>=0.1.6
anyio>=4.0.0
types-redis>=4.6.0
types-requests>=2.31.0
html2text>=2020.1.16
